@@ -1,15 +1,12 @@
-package com.example.android.corruption_whistleblower_app;
+package com.samgithiaka.corruption_whistleblower_app;
 
 import android.app.Activity;
 import android.app.ProgressDialog;
 import android.content.Context;
-import android.support.annotation.VisibleForTesting;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
 
 public class BaseActivity extends Activity {
-
-    @VisibleForTesting
     public ProgressDialog mProgressDialog;
 
     public void showProgressDialog() {
