@@ -1,9 +1,8 @@
-package com.example.android.corruption_whistleblower_app;
+package com.samgithiaka.corruption_whistleblower_app;
 
 import android.Manifest;
 import android.app.AlertDialog;
 import android.content.ClipboardManager;
-import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
@@ -14,8 +13,6 @@ import android.location.LocationManager;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.MediaStore;
-import android.support.annotation.NonNull;
-import androidx.core.app.ActivityCompat;
 import android.text.TextUtils;
 import android.util.Base64;
 import android.util.Log;
@@ -25,7 +22,10 @@ import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.example.android.corruption_whistleblower_app.Model.Post;
+import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
+
+import com.samgithiaka.corruption_whistleblower_app.Model.Post;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -38,32 +38,22 @@ import java.util.Calendar;
 import java.util.HashMap;
 import java.util.Map;
 
-import static com.firebase.ui.auth.AuthUI.TAG;
-
 public class reportActivity extends BaseActivity {
 
-    private final int PICK_IMAGE_REQUEST = 71;
-    private Uri filePath;
-    private TextView imageView;
-
-   private  double lon=36.825358;
-   private double lat=-1.291244;
-   private  String key;
-   private  String date;
-
     private static final String REQUIRED = "Required";
-
+    private final int PICK_IMAGE_REQUEST = 71;
+    private TextView imageView;
+    private double lon = 36.825358;
+    private double lat = -1.291244;
+    private String date;
     // [START declare_database_ref]
     private DatabaseReference mDatabase;
     // [END declare_database_ref]
 
     private EditText reportField;
     private EditText badgeNoField;
-    private Button location2;
     private EditText fullName;
     private EditText emailAddress;
-    private Button attachImage;
-    private Button submitButton;
     private EditText locationIncident;
     private String imageString = "";
 
@@ -80,11 +70,11 @@ public class reportActivity extends BaseActivity {
         reportField = findViewById(R.id.incidenceReport);
         badgeNoField = findViewById(R.id.badgeNo);
         imageView = findViewById(R.id.imageView);
-        location2 = findViewById(R.id.locationButton);
+        Button location2 = findViewById(R.id.locationButton);
         fullName = findViewById(R.id.fullName);
         emailAddress = findViewById(R.id.emailAddress);
-        submitButton = findViewById(R.id.submitButton);
-        attachImage = findViewById(R.id.buttonAttachImage);
+        Button submitButton = findViewById(R.id.submitButton);
+        Button attachImage = findViewById(R.id.buttonAttachImage);
         locationIncident = findViewById(R.id.locationIncident);
 
 
@@ -97,13 +87,13 @@ public class reportActivity extends BaseActivity {
 
                 // Acquire a reference to the system Location Manager
                 LocationManager locationManager =
-                        (LocationManager) reportActivity.this.getSystemService(Context.LOCATION_SERVICE);
+                        (LocationManager) reportActivity.this.getSystemService(LOCATION_SERVICE);
                 // Define a listener that responds to location updates
                 LocationListener locationListener = new LocationListener() {
                     public void onLocationChanged(Location location) {
                         // Called when a new location is found by the network location provider.
-                         lat = (location.getLatitude());
-                         lon = (location.getLongitude());
+                        lat = (location.getLatitude());
+                        lon = (location.getLongitude());
 
                     }
 
@@ -130,12 +120,12 @@ public class reportActivity extends BaseActivity {
 
         });
 
-        attachImage.setOnClickListener(new View.OnClickListener(){
+        attachImage.setOnClickListener(new View.OnClickListener() {
 
 
             //choosing the image
             @Override
-            public void onClick (View v){
+            public void onClick(View v) {
 
                 Intent intent = new Intent();
                 intent.setType("image/*");
@@ -162,8 +152,8 @@ public class reportActivity extends BaseActivity {
         final String location = locationIncident.getText().toString();
         final String name = fullName.getText().toString();
         final String email = emailAddress.getText().toString();
-        final String caseSerialNo="key";
-        final String date=Calendar.getInstance().getTime().toString();
+        final String caseSerialNo = "key";
+        final String date = Calendar.getInstance().getTime().toString();
         // Title is required
         if (TextUtils.isEmpty(report)) {
             reportField.setError(REQUIRED);
@@ -182,21 +172,21 @@ public class reportActivity extends BaseActivity {
         }
 
 
-
         mDatabase.child("Reports").addListenerForSingleValueEvent(
                 new ValueEventListener() {
 
-        @Override
-        public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                    @Override
+                    public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
 
-        writeNewPost(report, badgeNo, image, location, name, email,caseSerialNo,date);
+                        writeNewPost(report, badgeNo, image, location, name, email, caseSerialNo, date);
                         // Finish this Activity, back to the stream
                         setEditingEnabled(true);
 
                         // [END_EXCLUDE]
                     }
+
                     public void onCancelled(DatabaseError databaseError) {
-                        Log.w(TAG, "getUser:onCancelled", databaseError.toException());
+                        Log.w("hey", "getUser:onCancelled", databaseError.toException());
                         // [START_EXCLUDE]
                         setEditingEnabled(true);
                         // [END_EXCLUDE]
@@ -205,7 +195,7 @@ public class reportActivity extends BaseActivity {
         // [END single_value_read]
     }
 
-   private void setEditingEnabled(boolean enabled) {
+    private void setEditingEnabled(boolean enabled) {
         reportField.setEnabled(enabled);
         badgeNoField.setEnabled(enabled);
         imageView.setEnabled(enabled);
@@ -221,87 +211,82 @@ public class reportActivity extends BaseActivity {
 
 
     // [START write_fan_out]
-   private void writeNewPost( String report, String badgeNo, String image, String location, String name, String email, String caseSerialNo,String date) {
+    private void writeNewPost(String report, String badgeNo, String image, String location, String name, String email, String caseSerialNo, String date) {
 
-         key = mDatabase.push().getKey();
-        caseSerialNo=key;
+        String key = mDatabase.push().getKey();
+        caseSerialNo = key;
 
-        Post post = new Post(report, badgeNo, image, location, name, email,caseSerialNo,date);
+        Post post = new Post(report, badgeNo, image, location, name, email, caseSerialNo, date);
         Map<String, Object> postValues = post.toMap();
 
         Map<String, Object> childUpdates = new HashMap<>();
         childUpdates.put("" + key, postValues);
         mDatabase.child(key).setValue(post);
 
-     //Alert Dialog
+        //Alert Dialog
 
-       TextView showText = new TextView(this);
-       showText.setText("SUBMITED SUCCESFULLY.. LONG CLICK TO COPY TO CLIPBOARD AND SEND VIA CHAT CASE REF NO;" +key);
+        TextView showText = new TextView(this);
+        showText.setText("SUBMITED SUCCESFULLY.. LONG CLICK TO COPY TO CLIPBOARD AND SEND VIA CHAT CASE REF NO;" + key);
 // Add the Listener
-       showText.setOnLongClickListener(new View.OnLongClickListener() {
+        showText.setOnLongClickListener(new View.OnLongClickListener() {
 
-           @Override
-           public boolean onLongClick(View v) {
-               // Copy the Text to the clipboard
-               ClipboardManager manager =
-                       (ClipboardManager) getSystemService(CLIPBOARD_SERVICE);
-               TextView showTextParam = (TextView) v;
-               manager.setText( showTextParam.getText() );
-               // Show a message:
-               Toast.makeText(v.getContext(), "Text in clipboard",
-                       Toast.LENGTH_SHORT)
-                       .show();
-               return true;
-           }
-       });
-       AlertDialog.Builder dialog=new AlertDialog.Builder(this,R.style.AlertDialogStyle);
-       dialog.setView(showText);
-       dialog.setTitle("PLEASE COPY REF NO");
-       dialog.setPositiveButton("Ok",
-               new DialogInterface.OnClickListener() {
-                   public void onClick(DialogInterface dialog,
-                                       int which) {
-                       finish();
+            @Override
+            public boolean onLongClick(View v) {
+                // Copy the Text to the clipboard
+                ClipboardManager manager =
+                        (ClipboardManager) getSystemService(CLIPBOARD_SERVICE);
+                TextView showTextParam = (TextView) v;
+                manager.setText(showTextParam.getText());
+                // Show a message:
+                Toast.makeText(v.getContext(), "Text in clipboard",
+                        Toast.LENGTH_SHORT)
+                        .show();
+                return true;
+            }
+        });
+        AlertDialog.Builder dialog = new AlertDialog.Builder(this, R.style.AlertDialogStyle);
+        dialog.setView(showText);
+        dialog.setTitle("PLEASE COPY REF NO");
+        dialog.setPositiveButton("Ok",
+                new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog,
+                                        int which) {
+                        finish();
 
-                   }
-               });
-       AlertDialog alertDialog=dialog.create();
-       alertDialog.show();
+                    }
+                });
+        AlertDialog alertDialog = dialog.create();
+        alertDialog.show();
 
     }
     // [END write_fan_out]
 
 
-
-
-//set selected image filepath to string visible by user
+    //set selected image filepath to string visible by user
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
         imageView = findViewById(R.id.imageView);
-        if(requestCode == PICK_IMAGE_REQUEST && resultCode == RESULT_OK
-                && data != null && data.getData() != null )
-        {
-            filePath = data.getData();
+        if (requestCode == PICK_IMAGE_REQUEST && resultCode == RESULT_OK
+                && data != null && data.getData() != null) {
+            Uri filePath = data.getData();
             try {
                 Bitmap bitmap = MediaStore.Images.Media.getBitmap(getContentResolver(), filePath);
                 imageView.setText(filePath.toString());
                 //change the bitmap to string
-                imageString  = BitMapToString(bitmap);
-            }
-            catch (IOException e)
-            {
+                imageString = BitMapToString(bitmap);
+            } catch (IOException e) {
                 e.printStackTrace();
             }
         }
     }
 
     //convert bitmap to string
-    public String BitMapToString(Bitmap bitmap){
-        ByteArrayOutputStream baos=new  ByteArrayOutputStream();
-        bitmap.compress(Bitmap.CompressFormat.PNG,100, baos);
-        byte [] b=baos.toByteArray();
-        String temp= Base64.encodeToString(b, Base64.DEFAULT);
+    public String BitMapToString(Bitmap bitmap) {
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        bitmap.compress(Bitmap.CompressFormat.PNG, 100, baos);
+        byte[] b = baos.toByteArray();
+        String temp = Base64.encodeToString(b, Base64.DEFAULT);
         return temp;
     }
 
